+++ api/test/routes/user.test.js
@@ -10,7 +10,7 @@ const {
   POST_USER_ADDITIONAL_FIELDS,
   POST_USER_ONE_LESS_FIELD,
   POST_WRONG_USER_NO_ROLE,
-} = require('../mock-data/language-mock-data');
+} = require('../mock-data/user-mock-data');
 const { withAuthentication } = require('../utils/auth');
 
 /* 
@@ -44,8 +44,8 @@ describe('POST /user/ ', () => {
 
   test('API should create user', async () => {
     const body = POST_SIMPLE_USER;
-    const response = await request(app).post('/user/').send(body);
-
+    
+    const response = await withAuthentication(request(app).post('/user/').send(body));
     const message = response.body.message;
     const result = _.omit(response.body.result, ['_id', '__v']);
 
@@ -55,14 +55,14 @@ describe('POST /user/ ', () => {
 
   test('HTTP response status code', async () => {
     const body = POST_SIMPLE_USER;
-    const response = await request(app).post('/user/').send(body);
-
+    const response = await withAuthentication(request(app).post('/user/').send(body));
+    console.log(response)
     expect(response.status).toBe(200);
   });
 
   test('User Role is always 0', async () => {
     const body = POST_USER_ADMIN;
-    const response = await request(app).post('/user/').send(body);
+    const response = await withAuthentication(request(app).post('/user/').send(body));
 
     const message = response.body.message;
     const result = _.omit(response.body.result, ['_id', '__v']);
@@ -73,14 +73,14 @@ describe('POST /user/ ', () => {
 
   test('No id token results in error', async () => {
     const body = POST_WRONG_USER_NO_ID_TOKEN;
-    const response = await request(app).post('/user/').send(body);
+    const response = await withAuthentication(request(app).post('/user/').send(body));
 
     expect(response.status).toBeGreaterThanOrEqual(400);
   });
 
   test('No role still creates a new user', async () => {
     const body = POST_WRONG_USER_NO_ROLE;
-    const response = await request(app).post('/user/').send(body);
+    const response = await withAuthentication(request(app).post('/user/').send(body));
 
     const message = response.body.message;
     const result = _.omit(response.body.result, ['_id', '__v']);
@@ -91,7 +91,7 @@ describe('POST /user/ ', () => {
 
   test('Additional fields still create a new user', async () => {
     const body = POST_USER_ADDITIONAL_FIELDS;
-    const response = await request(app).post('/user/').send(body);
+    const response = await withAuthentication(request(app).post('/user/').send(body));
 
     const message = response.body.message;
     const result = _.omit(response.body.result, ['_id', '__v']);
@@ -102,7 +102,7 @@ describe('POST /user/ ', () => {
 
   test('One less field still create a new user', async () => {
     const body = POST_USER_ONE_LESS_FIELD;
-    const response = await request(app).post('/user/').send(body);
+    const response = await withAuthentication(request(app).post('/user/').send(body));
 
     const message = response.body.message;
     const result = _.omit(response.body.result, ['_id', '__v']);
